feat: switch from pyright to pylsp
Pyright is unable to resolve some imports and types when using Django.
This is pretty annoying. Pylsp seems to be doing a better job here.
Pylsp comes with a bunch of plugins that do linting and formatting. I
don't need that since I am using ruff. Therefore, they are disabled.
Autoimport sounds useful. It takes some time to load however for some
files. Let's see if it stays.